+++ Backend/src/main/java/com/sn/budgetbee/services/ExitServiceImpl.java
@@ -108,12 +108,12 @@ public class ExitServiceImpl implements ExitService{
     }
 
     @Override
-    public Integer exitByMonth(Integer id, String month) {
+    public  Double exitByMonth(Integer id, String month) {
         return EXIT_DAO.findTotalExitByMonth(id, month);
     }
 
     @Override
-    public Integer exitByYear(Integer id, String year) {
+    public  Double exitByYear(Integer id, String year) {
         return  EXIT_DAO.findTotalExitByYear(id, year);
     }
 }
